Now using Time.deltaTime
The MenuUtility class changes the timeScale to 0 to freeze the game on pause

+++ Accropoly/Assets/0_Scripts/Authoring+Monobehaviour/TileAgeingConfigAuthoring.cs
@@ -30,6 +30,6 @@ public struct TileAgeingConfig : IComponentData
 {
     public uint seed;
     public int2 randomAgeRange;
-    public int maxAge;
+    public float maxAge;
     public TileType newTileType;
 }
