+++ Accropoly/Assets/0_Scripts/Components/AgingTile.cs
@@ -1,9 +1,6 @@
-using System.Collections;
-using System.Collections.Generic;
 using Unity.Entities;
-using UnityEngine;
 
 public struct AgingTile : IComponentData
 {
-    public int age;
+    public float age;
 }
